WDFO-81 - create exceptions module
